리뷰 1개를 생성해주는 역할을 ReviewPanel 별도의 클래스로 분리시킴
