+++ reviewsdata.csv
@@ -1,6 +1,5 @@
-봄이1,1234,객체 지향 어려워요,어려운데 재밌는거... 맞나요?,DELETED
+봄이1,1234,객체 지향 어려워요,어려운데 재밌는거... 맞나요?,DISPLAY
 봄이2,abcd,클린 코드 뭐지~,다 읽으셨나요? 저는 다 읽었어요 ㅎㅎㅎ :),DELETED
-봄이3,클린코드,프로그래머 미안,북스터디 해요~,DELETED
-봄이3,1234,오늘은 뭘 읽을까!,리뷰는 짧을 수록 좋은것이다~,DISPLAY
-봄이 4,1234,안녕하세요~,100자 이내로 리뷰를 적어주세요 100자 이내로 적으라는데 저는 그럴 생각이 없거든요~ 아주많이많이 적을거예요 어떻게 생각하세요? 근데 이거 계속 적다보면 레이아웃 시망 진짜 어쩔ㅠ,DISPLAY
-봄이 4,1234,안녕하세요~,100자 이내로 리뷰를 적어주세요 100자 이내로 적으라는데 저는 그럴 생각이 없거든요~ 아주많이많이 적을거예요 어떻게 생각하세요? 근데 이거 계속 적다보면 레이아웃 시망 진짜 어쩔ㅠ,DELETED
+봄이3,클린코드,프로그래머 미안 어려워...;,북스터디 해요~,DISPLAY
+봄이4,1234,오늘은 뭘 읽을까!,리뷰는 짧을 수록 좋은것이다~,DISPLAY
+봄이 5,1234,제목은 뭐라고 지을까요~,100자 이내로 리뷰를 적어주세요 100자 넘어가면 자동으로 사라지는거 실화냐?!,DISPLAY
